#127: Review comments. Typo nitpicks, missing final, removing duplication.
