+++ core/src/main/java/io/parsingdata/metal/token/Cho.java
@@ -29,7 +29,7 @@ import io.parsingdata.metal.encoding.Encoding;
 
 public class Cho extends Token {
 
-    private final Token[] tokens; // Private because array contents is mutable.
+    private final Token[] tokens; // Private because array content is mutable.
 
     public Cho(final String name, final Encoding encoding, final Token... tokens) {
         super(name, encoding);
