+++ core/src/test/java/io/parsingdata/metal/token/TieTest.java
@@ -39,6 +39,7 @@ import static io.parsingdata.metal.Shorthand.rev;
 import static io.parsingdata.metal.Shorthand.seq;
 import static io.parsingdata.metal.Shorthand.sub;
 import static io.parsingdata.metal.Shorthand.tie;
+import static io.parsingdata.metal.Util.inflate;
 import static io.parsingdata.metal.data.ConstantSlice.create;
 import static io.parsingdata.metal.data.selection.ByName.getAllValues;
 import static io.parsingdata.metal.data.selection.ByType.getReferences;
@@ -173,32 +174,6 @@ public class TieTest {
         output[3] = (byte) (data.length & 0xff);
         System.arraycopy(data, 0, output, 4, data.length);
         return output;
-    }
-
-    public static ValueExpression inflate(final ValueExpression target) {
-        return new UnaryValueExpression(target) {
-            @Override
-            public OptionalValue eval(final Value value, final Environment environment, final Encoding encoding) {
-                return value.operation(new ValueOperation() {
-                    @Override
-                    public OptionalValue execute(final Value value) {
-                        final Inflater inf = new Inflater(true);
-                        inf.setInput(value.getValue());
-                        final byte[] dataReceiver = new byte[512];
-                        final ByteArrayOutputStream out = new ByteArrayOutputStream();
-                        while(!inf.finished()) {
-                            try {
-                                final int processed = inf.inflate(dataReceiver);
-                                out.write(dataReceiver, 0, processed);
-                            } catch (final DataFormatException e) {
-                                return OptionalValue.empty();
-                            }
-                        }
-                        return OptionalValue.of(new Value(create(out.toByteArray()), encoding));
-                    }
-                });
-            }
-        };
     }
 
     @Test
