+++ formats/src/main/java/io/parsingdata/metal/format/Callback.java
@@ -18,10 +18,7 @@ package io.parsingdata.metal.format;
 
 import static io.parsingdata.metal.data.ConstantSlice.create;
 
-import java.io.ByteArrayOutputStream;
 import java.util.zip.CRC32;
-import java.util.zip.DataFormatException;
-import java.util.zip.Inflater;
 
 import io.parsingdata.metal.data.Environment;
 import io.parsingdata.metal.encoding.Encoding;
@@ -49,32 +46,6 @@ public final class Callback {
                                                                                                              (byte)((crcValue & 0xff0000) >> 16),
                                                                                                              (byte)((crcValue & 0xff00) >> 8),
                                                                                                              (byte) (crcValue & 0xff) })), encoding));
-                    }
-                });
-            }
-        };
-    }
-
-    public static ValueExpression inflate(final ValueExpression target) {
-        return new UnaryValueExpression(target) {
-            @Override
-            public OptionalValue eval(final Value value, final Environment environment, final Encoding encoding) {
-                return value.operation(new ValueOperation() {
-                    @Override
-                    public OptionalValue execute(final Value value) {
-                        final Inflater inf = new Inflater(true);
-                        inf.setInput(value.getValue());
-                        final byte[] dataReceiver = new byte[512];
-                        final ByteArrayOutputStream out = new ByteArrayOutputStream();
-                        while(!inf.finished()) {
-                            try {
-                                final int processed = inf.inflate(dataReceiver);
-                                out.write(dataReceiver, 0, processed);
-                            } catch (final DataFormatException e) {
-                                return OptionalValue.empty();
-                            }
-                        }
-                        return OptionalValue.of(new Value(create(out.toByteArray()), encoding));
                     }
                 });
             }
