+++ formats/src/main/java/io/parsingdata/metal/format/PNG.java
@@ -16,11 +16,20 @@
 
 package io.parsingdata.metal.format;
 
+import static io.parsingdata.metal.Shorthand.cat;
+import static io.parsingdata.metal.Shorthand.con;
+import static io.parsingdata.metal.Shorthand.def;
+import static io.parsingdata.metal.Shorthand.eq;
+import static io.parsingdata.metal.Shorthand.eqNum;
+import static io.parsingdata.metal.Shorthand.last;
+import static io.parsingdata.metal.Shorthand.not;
+import static io.parsingdata.metal.Shorthand.ref;
+import static io.parsingdata.metal.Shorthand.rep;
+import static io.parsingdata.metal.Shorthand.seq;
+import static io.parsingdata.metal.format.Callback.crc32;
+
 import io.parsingdata.metal.encoding.Encoding;
 import io.parsingdata.metal.token.Token;
-
-import static io.parsingdata.metal.Shorthand.*;
-import static io.parsingdata.metal.format.Callback.crc32;
 
 public final class PNG {
