+++ formats/src/main/java/io/parsingdata/metal/format/ZIP.java
@@ -16,14 +16,28 @@
 
 package io.parsingdata.metal.format;
 
+import static io.parsingdata.metal.Shorthand.cho;
+import static io.parsingdata.metal.Shorthand.con;
+import static io.parsingdata.metal.Shorthand.count;
+import static io.parsingdata.metal.Shorthand.def;
+import static io.parsingdata.metal.Shorthand.eq;
+import static io.parsingdata.metal.Shorthand.eqNum;
+import static io.parsingdata.metal.Shorthand.expTrue;
+import static io.parsingdata.metal.Shorthand.first;
+import static io.parsingdata.metal.Shorthand.last;
+import static io.parsingdata.metal.Shorthand.offset;
+import static io.parsingdata.metal.Shorthand.ref;
+import static io.parsingdata.metal.Shorthand.rep;
+import static io.parsingdata.metal.Shorthand.self;
+import static io.parsingdata.metal.Shorthand.seq;
+import static io.parsingdata.metal.Shorthand.sub;
+import static io.parsingdata.metal.Util.inflate;
+import static io.parsingdata.metal.format.Callback.crc32;
+
 import io.parsingdata.metal.encoding.ByteOrder;
 import io.parsingdata.metal.encoding.Encoding;
 import io.parsingdata.metal.expression.Expression;
 import io.parsingdata.metal.token.Token;
-
-import static io.parsingdata.metal.Shorthand.*;
-import static io.parsingdata.metal.format.Callback.crc32;
-import static io.parsingdata.metal.format.Callback.inflate;
 
 /*
  * Implements limited subset of ZIP file format:
